+++ src/types.ts
@@ -318,6 +318,9 @@ export type Router<Endpoints extends RouteEndpoint[]> = GetHttpHandlers<Endpoint
 
 export type InferEndpoints<T> = T extends Router<infer E> ? E : never
 
+/** @experimental */
+export type FetchLike = (input: RequestInfo | URL, init?: RequestInit) => Promise<Response>
+
 export interface ClientOptions extends Pick<RequestInit, "cache" | "credentials" | "mode"> {
     /**
      * Base URL for the router client to make requests to the server.
@@ -334,6 +337,11 @@ export interface ClientOptions extends Pick<RequestInit, "cache" | "credentials"
      * Default headers to include in every request made by the client.
      */
     headers?: RequestHeaders | (() => RequestHeaders | Promise<RequestHeaders>)
+    /**
+     * @experimental
+     * Custom fetch function to be used by the client instead of the global fetch.
+     */
+    fetch?: FetchLike
 }
 
 export interface RequestHeaders extends Record<string, number | string | string[] | undefined> {