+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Per-package version history is maintained inside each package’s own `CHANGELOG
 
 ### Added
 
+- Added a `fetch` configuration option in `createClient` to pass a custom fetch implementation. This option can be overridden by JavaScript library or framework integrations. [#40](https://github.com/aura-stack-ts/router/pull/40)
+
 - Added `cache`, `credentials`, and `mode` RequestInit options in `createClient`. These options are passed to all requests made by the client. [#37](https://github.com/aura-stack-ts/router/pull/37)
 
 - Added support for multiple HTTP methods in endpoint definitions via `createEndpoint`. Multiple HTTP methods can now match the same route. [#36](https://github.com/aura-stack-ts/router/pull/36)